+++ app/components/bulma/title_component.rb
@@ -2,15 +2,16 @@
 
 module Bulma
   class TitleComponent < Component
-    def initialize(as: :h1, size: 1)
+    def initialize(title = nil, as: :h1, size: 1)
       raise ArgumentError, "invalid value to size" if size > 6
 
+      @title = title
       @as = as
       @size = size
     end
 
     def call
-      content_tag @as, content, class: class_names(
+      content_tag @as, @title || content, class: class_names(
         "title",
         "is-#{@size}"
       )